LayoutTests/canvas/philip/tests/2d.composite.uncovered.nocontext.copy-expected.txt

< [index] >
2d.composite.uncovered.nocontext.copy
drawImage() of a canvas with no context draws pixels as (0,0,0,0), and does not leave the pixels unchanged.
References: 2d.composite.copy
Actual output:
Expected output:

Passed
